Move SoDatumLabel from Sketcher to Core so that it can be reused anywhere.
